+++ partiql-spi/src/main/kotlin/org/partiql/spi/function/builtins/FnTimes.kt
@@ -6,8 +6,10 @@ package org.partiql.spi.function.builtins
 import org.partiql.spi.function.Fn
 import org.partiql.spi.function.builtins.internal.PErrors
 import org.partiql.spi.types.PType
+import org.partiql.spi.utils.IntervalUtils
 import org.partiql.spi.utils.NumberUtils.byteOverflows
 import org.partiql.spi.utils.NumberUtils.shortOverflows
+import org.partiql.spi.utils.getNumber
 import org.partiql.spi.value.Datum
 
 internal object FnTimes : DiadicArithmeticOperator("times") {
@@ -114,6 +116,24 @@ internal object FnTimes : DiadicArithmeticOperator("times") {
             val arg0 = args[0].double
             val arg1 = args[1].double
             Datum.doublePrecision(arg0 * arg1)
+        }
+    }
+
+    override fun getIntervalNumberInstance(lhs: PType, rhs: PType): Fn {
+        val op = IntervalUtils.intervalMultiply(lhs)
+        return basic(lhs, lhs, rhs) { args ->
+            val interval = args[0]
+            val number = args[1].getNumber()
+            op(interval, number)
+        }
+    }
+
+    override fun getNumberIntervalInstance(lhs: PType, rhs: PType): Fn {
+        val op = IntervalUtils.intervalMultiply(rhs)
+        return basic(rhs, lhs, rhs) { args ->
+            val number = args[0].getNumber()
+            val interval = args[1]
+            op(interval, number)
         }
     }
 }