+++ CHANGELOG.md
@@ -49,6 +49,7 @@ rewritten to their query representation
 - CLI printing of negative interval literals
 - Fix alias collision problem when multiple table references need auto-generated aliases
 - Display of `Ion` in PartiQL CLI output
+- Support `INTERVAL` times and divide with numerics
 
 ### Removed